+++ src/js_sync.hpp
@@ -996,6 +996,12 @@ void SyncClass<T>::reconnect(ContextType ctx, ObjectType this_object, Arguments 
 }
 
 template<typename T>
+void SyncClass<T>::has_existing_sessions(ContextType ctx, ObjectType this_object, Arguments &args, ReturnValue & return_value) {
+    args.validate_count(0);
+    return_value.set(syncManagerShared<T>(ctx).has_existing_sessions());
+}
+
+template<typename T>
 std::function<SyncBindSessionHandler> SyncClass<T>::session_bind_callback(ContextType ctx, ObjectType sync_constructor)
 {
     Protected<typename T::GlobalContext> protected_ctx(Context<T>::get_global_context(ctx));
@@ -1102,6 +1108,22 @@ void SyncClass<T>::populate_sync_config(ContextType ctx, ObjectType realm_constr
         config.sync_config->error_handler = std::move(error_handler);
         config.sync_config->is_partial = is_partial;
 
+        SyncSessionStopPolicy session_stop_policy = SyncSessionStopPolicy::AfterChangesUploaded;
+        ValueType session_stop_policy_value = Object::get_property(ctx, sync_config_object, "_sessionStopPolicy");
+        if (!Value::is_undefined(ctx, session_stop_policy_value)) {
+            std::string stop_session = Value::validated_to_string(ctx, session_stop_policy_value, "_sessionStopPolicy");
+            if (stop_session == std::string("immediately")) {
+                session_stop_policy = SyncSessionStopPolicy::Immediately;
+            } else if (stop_session == std::string("never")) {
+                session_stop_policy = SyncSessionStopPolicy::LiveIndefinitely;
+            } else if (stop_session == std::string("after-upload")) {
+                session_stop_policy = SyncSessionStopPolicy::AfterChangesUploaded;
+            } else {
+                throw std::invalid_argument("Unknown argument for _sessionStopPolicy: " + stop_session);
+            }
+        }
+        config.sync_config->stop_policy = session_stop_policy;
+
         ValueType custom_partial_sync_identifier_value = Object::get_property(ctx, sync_config_object, "customQueryBasedSyncIdentifier");
         if (!Value::is_undefined(ctx, custom_partial_sync_identifier_value)) {
             config.sync_config->custom_partial_sync_identifier = std::string(Value::validated_to_string(ctx, custom_partial_sync_identifier_value, "customQueryBasedSyncIdentifier"));
@@ -1190,10 +1212,6 @@ void SyncClass<T>::populate_sync_config_for_ssl(ContextType ctx, ObjectType conf
     }
 }
 
-template<typename T>
-void SyncClass<T>::has_existing_sessions(ContextType ctx, ObjectType this_object, Arguments &args, ReturnValue &return_value) {
-    return_value.set(SyncManager::shared().has_existing_sessions());
-}
 
 } // js
 } // realm